(pd) misc light refactoring & cleanup
* document `ingredientsForContainer` selector in README
* factor out IngredientSelectionModal from Deck to its own component
* move a few utility fns into utils.js
* update/remove old TODOs and comments
